Execption handled and added unit tests

+++ src/main/java/com/spring/movlib/controller/MovieController.java
@@ -12,15 +12,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.spring.movlib.model.CategoryName;
 import com.spring.movlib.model.Movie;
@@ -41,37 +40,49 @@ public class MovieController {
 
 	@PostMapping("/movie")
 	Movie addMovie(@Valid @RequestBody Movie movie) throws ValidationException {
+		if(movie.getId() != 0)
+			throw new ValidationException("Id can not be pre-defined");
 		return movieService.addMovie(movie);
 	}
 	 
 
 	@GetMapping("/movie")
 	Iterable<Movie> listAllMovies() {
-		return movieService.read();
+		Iterable<Movie> foundMovies = movieService.read();
+		if(foundMovies == null)
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND,"Movies not Found");
+		return foundMovies;
 	}
 	
-	@GetMapping("/category/{categoryName}")
-	Iterable<Movie> listMoviesByCategory(@PathVariable CategoryName categoryName) {
-		return categoryService.findMoviesByCategory(categoryName);
-	}
-
 	@PutMapping("/movie")
 	ResponseEntity<Movie> updateMovie(@RequestBody Movie movie) {
 		Movie updatedMovie = movieService.update(movie);
-		if ( updatedMovie != null)
-			return new ResponseEntity<Movie>(updatedMovie, HttpStatus.OK);
-		else
-			return new ResponseEntity<Movie>(movie, HttpStatus.BAD_REQUEST);
+		if ( updatedMovie == null)
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid Movie Request/Movie Not Present");
+		return new ResponseEntity<Movie>(updatedMovie, HttpStatus.OK);
 	}
-
+	
 	@DeleteMapping("/movie/{id}")
 	void delete(@PathVariable Integer id) {
+		if(movieService.findMovieById(id).isEmpty())
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND,"Movie not Found");
 		movieService.deleteMovieById(id);
 	}
 
 	@GetMapping("/movie/{id}")
 	Optional<Movie> findById(@PathVariable Integer id) {
-		return movieService.findMovieById(id);
+		Optional<Movie> foundMovie = movieService.findMovieById(id);
+		if(foundMovie.isEmpty())
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND,"Movie not Found");
+		return foundMovie;
+	}
+	
+	@GetMapping("/category/{categoryName}")
+	Iterable<Movie> listMoviesByCategory(@PathVariable CategoryName categoryName) {
+		Iterable<Movie> foundMovies = categoryService.findMoviesByCategory(categoryName);
+		if(foundMovies == null)
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND,"No Movies Found for the Category: "+ categoryName);
+		return foundMovies;
 	}
 
 	@GetMapping("/movie/search")
@@ -89,12 +100,6 @@ public class MovieController {
 			return movieService.findMoviesByDirector(director);
 		else
 			return movieService.findAllMovies();
-	}
-
-	@ResponseStatus(HttpStatus.BAD_REQUEST)
-	@ExceptionHandler(ValidationException.class)
-	String exceptionHandler(ValidationException e) {
-		return e.getMessage();
 	}
 
 }
